[Issue #174] changed product to product type. draft the product costs draft the product
