+++ koalixcrm/crm/factories/factory_product.py
@@ -2,7 +2,7 @@
 
 import factory
 import datetime
-from koalixcrm.crm.models import Product
+from koalixcrm.crm.models import ProductType
 from koalixcrm.crm.factories.factory_unit import StandardUnitFactory
 from koalixcrm.crm.factories.factory_user import StaffUserFactory
 from koalixcrm.accounting.factories.factory_product_category import StandardProductCategoryFactory
@@ -11,7 +11,7 @@ from koalixcrm.test_support_functions import make_date_utc
 
 class StandardProductFactory(factory.django.DjangoModelFactory):
     class Meta:
-        model = Product
+        model = ProductType
         django_get_or_create = ('product_number',)
 
     title = "This is a test Product"
